Update Annual_SST_cycle_indicator.Rmd
Ln #27 - I added the lab location as suggested by the OI SST website citation.  
Ln #175 - We might want to explain why the long-term mean is just between 1982 and 2012 and not 2017.  If that is the standard practice, then we need to provide the reference.
The figure caption should indicate that the "long-term mean" and STD are for the 1982-2012 time period and explain what the orange and cyan colors represent (in addition to the red and blue).